Move suggest_ref_mut into rustc_mir::borrow_check
